Extract types to a separate file
